refactor: share internal resolver util
